fea: fix test + improved tests on the acountability side
