+++ backend/tests/test_penas_controller.py
@@ -40,17 +40,15 @@ from core.application.queries.pena_queries import (
     ListPenasForAdminQuery,
     ListPenasForUserQuery,
 )
-from core.application.use_cases.manage_pena_accountability_usecase import (
+from core.domain.errors import (
+    InvalidLinkTokenError,
     InvalidPenaAccountabilityDataError,
+    InvalidPenaLabelsDataError,
+    InvalidProfileImageError,
     PenaAccountabilityAccessDeniedError,
     PenaAccountabilityExpenseNotFoundError,
     PenaAccountabilityMemberNotFoundError,
     PenaAccountabilityPenaNotFoundError,
-)
-from core.domain.errors import (
-    InvalidLinkTokenError,
-    InvalidPenaLabelsDataError,
-    InvalidProfileImageError,
     PenaLabelsAccessDeniedError,
     PenaLabelsPenaNotFoundError,
     PenaLinkAccessDeniedError,
